+++ src/main/kotlin/net/plpgsql/ideadebugger/service/PlProcessWatcher.kt
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2022. Alexandre Boyer
+ */
+
+package net.plpgsql.ideadebugger.service
+
+import net.plpgsql.ideadebugger.run.PlProcess
+
+interface PlProcessWatcher {
+    fun isDebugging(): Boolean
+    fun processStarted(process: PlProcess)
+    fun processFinished(process: PlProcess)
+}