+++ src/main/kotlin/net/plpgsql/ideadebugger/vfs/PlVirtualFileSystem.kt
@@ -2,14 +2,13 @@
  * Copyright (c) 2022. Alexandre Boyer
  */
 
-package net.plpgsql.ideadebugger
+package net.plpgsql.ideadebugger.vfs
 
 import com.intellij.openapi.application.runInEdt
 import com.intellij.openapi.application.runWriteAction
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.VirtualFileManager
 import com.intellij.openapi.vfs.ex.dummy.DummyCachingFileSystem
-import com.intellij.testFramework.runInEdtAndGet
 import java.util.*